Move custom range storage classes to the dashboard package
