+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/dashboard/data/StatsCustomDateRangeDataStore.kt
@@ -1,8 +1,9 @@
-package com.woocommerce.android.ui.mystore.data
+package com.woocommerce.android.ui.dashboard.data
 
 import androidx.datastore.core.DataStore
 import com.woocommerce.android.datastore.DataStoreQualifier
 import com.woocommerce.android.datastore.DataStoreType
+import com.woocommerce.android.ui.mystore.data.CustomDateRange
 import javax.inject.Inject
 
 class StatsCustomDateRangeDataStore @Inject constructor(
